refactor: consolidate opcli concierge into install + provision
- Move snap installation to `opcli install concierge` (consistent with
  `opcli install spread` and `opcli install tox`)
- Rename `opcli provision run` → `opcli provision prepare` with -c flag
  for concierge file path
- Remove sudo from provision_prepare (caller handles privileges)
- Remove `opcli concierge` command group entirely
- Update spread backend scripts to use the new commands
- Update docs (spec + divergences)

Closes #144
