+++ src/opcli/core/spread.py
@@ -282,7 +282,8 @@ else
 fi
 opcli install spread
 opcli install tox
-opcli concierge prepare -c "$CONCIERGE"
+opcli install concierge
+opcli provision prepare -c "$CONCIERGE"
 runuser -l ubuntu -c \
   "cd \\"${SPREAD_PATH}\\" && opcli provision registry"
 opcli provision load
@@ -306,7 +307,8 @@ else
 fi
 opcli install spread
 opcli install tox
-opcli concierge prepare -c "$CONCIERGE"
+opcli install concierge
+opcli provision prepare -c "$CONCIERGE"
 """
 
 _CI_PREPARE_AFTER_USER = """\
